requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
Click==8.1.3
itsdangerous==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46 
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.4
python-dotenv==0.21.0
gunicorn==20.1.0
mysql-connector-python==8.0.30